#11 Added StandardEmptyScreen widget
Using the mainAxisAlignment field for the Column results in it being already centered, so we don't need to wrap the entire Column with a Center widget.

Modifying pubspec.yaml so that it only has assets/ as assets field results in the images not loading, I have to list all the folders in the assets folder.